Weighted CRF: moved classes to new module
Simplified Lannoy implementation.
