--- allennlp/modules/conditional_random_field_lannoy.py
@@ -1,486 +1,0 @@
-"""
-Conditional random field
-"""
-from typing import List, Tuple, Dict, Union
-
-import torch
-
-from allennlp.common.checks import ConfigurationError
-import allennlp.nn.util as util
-
-VITERBI_DECODING = Tuple[List[int], float]  # a list of tags, and a viterbi score
-
-
-def allowed_transitions(constraint_type: str, labels: Dict[int, str]) -> List[Tuple[int, int]]:
-    """
-    Given labels and a constraint type, returns the allowed transitions. It will
-    additionally include transitions for the start and end states, which are used
-    by the conditional random field.
-
-    # Parameters
-
-    constraint_type : `str`, required
-        Indicates which constraint to apply. Current choices are
-        "BIO", "IOB1", "BIOUL", and "BMES".
-    labels : `Dict[int, str]`, required
-        A mapping {label_id -> label}. Most commonly this would be the value from
-        Vocabulary.get_index_to_token_vocabulary()
-
-    # Returns
-
-    `List[Tuple[int, int]]`
-        The allowed transitions (from_label_id, to_label_id).
-    """
-    num_labels = len(labels)
-    start_tag = num_labels
-    end_tag = num_labels + 1
-    labels_with_boundaries = list(labels.items()) + [(start_tag, "START"), (end_tag, "END")]
-
-    allowed = []
-    for from_label_index, from_label in labels_with_boundaries:
-        if from_label in ("START", "END"):
-            from_tag = from_label
-            from_entity = ""
-        else:
-            from_tag = from_label[0]
-            from_entity = from_label[1:]
-        for to_label_index, to_label in labels_with_boundaries:
-            if to_label in ("START", "END"):
-                to_tag = to_label
-                to_entity = ""
-            else:
-                to_tag = to_label[0]
-                to_entity = to_label[1:]
-            if is_transition_allowed(constraint_type, from_tag, from_entity, to_tag, to_entity):
-                allowed.append((from_label_index, to_label_index))
-    return allowed
-
-
-def is_transition_allowed(
-    constraint_type: str, from_tag: str, from_entity: str, to_tag: str, to_entity: str
-):
-    """
-    Given a constraint type and strings `from_tag` and `to_tag` that
-    represent the origin and destination of the transition, return whether
-    the transition is allowed under the given constraint type.
-
-    # Parameters
-
-    constraint_type : `str`, required
-        Indicates which constraint to apply. Current choices are
-        "BIO", "IOB1", "BIOUL", and "BMES".
-    from_tag : `str`, required
-        The tag that the transition originates from. For example, if the
-        label is `I-PER`, the `from_tag` is `I`.
-    from_entity : `str`, required
-        The entity corresponding to the `from_tag`. For example, if the
-        label is `I-PER`, the `from_entity` is `PER`.
-    to_tag : `str`, required
-        The tag that the transition leads to. For example, if the
-        label is `I-PER`, the `to_tag` is `I`.
-    to_entity : `str`, required
-        The entity corresponding to the `to_tag`. For example, if the
-        label is `I-PER`, the `to_entity` is `PER`.
-
-    # Returns
-
-    `bool`
-        Whether the transition is allowed under the given `constraint_type`.
-    """
-
-    if to_tag == "START" or from_tag == "END":
-        # Cannot transition into START or from END
-        return False
-
-    if constraint_type == "BIOUL":
-        if from_tag == "START":
-            return to_tag in ("O", "B", "U")
-        if to_tag == "END":
-            return from_tag in ("O", "L", "U")
-        return any(
-            [
-                # O can transition to O, B-* or U-*
-                # L-x can transition to O, B-*, or U-*
-                # U-x can transition to O, B-*, or U-*
-                from_tag in ("O", "L", "U") and to_tag in ("O", "B", "U"),
-                # B-x can only transition to I-x or L-x
-                # I-x can only transition to I-x or L-x
-                from_tag in ("B", "I") and to_tag in ("I", "L") and from_entity == to_entity,
-            ]
-        )
-    elif constraint_type == "BIO":
-        if from_tag == "START":
-            return to_tag in ("O", "B")
-        if to_tag == "END":
-            return from_tag in ("O", "B", "I")
-        return any(
-            [
-                # Can always transition to O or B-x
-                to_tag in ("O", "B"),
-                # Can only transition to I-x from B-x or I-x
-                to_tag == "I" and from_tag in ("B", "I") and from_entity == to_entity,
-            ]
-        )
-    elif constraint_type == "IOB1":
-        if from_tag == "START":
-            return to_tag in ("O", "I")
-        if to_tag == "END":
-            return from_tag in ("O", "B", "I")
-        return any(
-            [
-                # Can always transition to O or I-x
-                to_tag in ("O", "I"),
-                # Can only transition to B-x from B-x or I-x, where
-                # x is the same tag.
-                to_tag == "B" and from_tag in ("B", "I") and from_entity == to_entity,
-            ]
-        )
-    elif constraint_type == "BMES":
-        if from_tag == "START":
-            return to_tag in ("B", "S")
-        if to_tag == "END":
-            return from_tag in ("E", "S")
-        return any(
-            [
-                # Can only transition to B or S from E or S.
-                to_tag in ("B", "S") and from_tag in ("E", "S"),
-                # Can only transition to M-x from B-x, where
-                # x is the same tag.
-                to_tag == "M" and from_tag in ("B", "M") and from_entity == to_entity,
-                # Can only transition to E-x from B-x or M-x, where
-                # x is the same tag.
-                to_tag == "E" and from_tag in ("B", "M") and from_entity == to_entity,
-            ]
-        )
-    else:
-        raise ConfigurationError(f"Unknown constraint type: {constraint_type}")
-
-
-class ConditionalRandomFieldLannoy(torch.nn.Module):
-    """
-    This module uses the "forward-backward" algorithm to compute
-    the log-likelihood of its inputs assuming a conditional random field model.
-
-    See, e.g. http://www.cs.columbia.edu/~mcollins/fb.pdf
-
-    # Parameters
-
-    num_tags : `int`, required
-        The number of tags.
-    label_weights : `List[float]`, optional (default=`None`)
-        A list of weights to be used in the loss function in order to
-        give different weights for each token depending on its label.
-        `len(label_weights)` must be equal to `num_tags`. This is useful to
-        deal with highly unbalanced datasets. The method implemented here was based on
-        the paper *Weighted conditional random fields for supervised interpatient heartbeat
-        classification* proposed by De Lannoy et. al (2019).
-        See https://perso.uclouvain.be/michel.verleysen/papers/ieeetbe12gdl.pdf
-    constraints : `List[Tuple[int, int]]`, optional (default = `None`)
-        An optional list of allowed transitions (from_tag_id, to_tag_id).
-        These are applied to `viterbi_tags()` but do not affect `forward()`.
-        These should be derived from `allowed_transitions` so that the
-        start and end transitions are handled correctly for your tag type.
-    include_start_end_transitions : `bool`, optional (default = `True`)
-        Whether to include the start and end transition parameters.
-    """
-
-    def __init__(
-        self,
-        num_tags: int,
-        label_weights: List[float],
-        constraints: List[Tuple[int, int]] = None,
-        include_start_end_transitions: bool = True,
-    ) -> None:
-        super().__init__()
-        self.num_tags = num_tags
-
-        # transitions[i, j] is the logit for transitioning from state i to state j.
-        self.transitions = torch.nn.Parameter(torch.empty(num_tags, num_tags))
-
-        # _constraint_mask indicates valid transitions (based on supplied constraints).
-        # Include special start of sequence (num_tags + 1) and end of sequence tags (num_tags + 2)
-        if constraints is None:
-            # All transitions are valid.
-            constraint_mask = torch.full((num_tags + 2, num_tags + 2), 1.0)
-        else:
-            constraint_mask = torch.full((num_tags + 2, num_tags + 2), 0.0)
-            for i, j in constraints:
-                constraint_mask[i, j] = 1.0
-
-        self._constraint_mask = torch.nn.Parameter(constraint_mask, requires_grad=False)
-
-        # Also need logits for transitioning from "start" state and to "end" state.
-        self.include_start_end_transitions = include_start_end_transitions
-        if include_start_end_transitions:
-            self.start_transitions = torch.nn.Parameter(torch.Tensor(num_tags))
-            self.end_transitions = torch.nn.Parameter(torch.Tensor(num_tags))
-
-        # If label_weights is not given, use 1.0 for all weights.
-        if label_weights is None:
-            label_weights = [1.0] * num_tags
-        self.label_weights = torch.nn.Parameter(torch.Tensor(label_weights), requires_grad=False)
-
-        self.reset_parameters()
-
-    def reset_parameters(self):
-        torch.nn.init.xavier_normal_(self.transitions)
-        if self.include_start_end_transitions:
-            torch.nn.init.normal_(self.start_transitions)
-            torch.nn.init.normal_(self.end_transitions)
-
-    def _input_likelihood(
-        self, logits: torch.Tensor, tags: torch.Tensor, mask: torch.BoolTensor
-    ) -> torch.Tensor:
-        """
-        Computes the (batch_size,) denominator term for the log-likelihood, which is the
-        sum of the likelihoods across all possible state sequences.
-
-        Compute this value using the scaling trick instead of the log domain trick, since
-        this is necessary to implement the label-weighting method by Lannoy et al. (2012).
-        """
-        batch_size, sequence_length, num_tags = logits.size()
-
-        # Transpose batch size and sequence dimensions
-        mask = mask.transpose(0, 1).contiguous()
-        logits = logits.transpose(0, 1).contiguous()
-        tags = tags.transpose(0, 1).contiguous()
-
-        # insert an 1-sized second dimension to match z.shape
-        label_weights = self.label_weights.view(num_tags, 1)
-
-        # emit_scores.shape = (batch_size, num_tags)
-        emit_scores = logits[0]
-
-        # Initial alpha is the (batch_size, num_tags) tensor of likelihoods combining the
-        # transitions to the initial states and the logits for the first timestep.
-        # alpha.shape = (batch_size, num_tags)
-        if self.include_start_end_transitions:
-            alpha = torch.exp(self.start_transitions.view(1, num_tags) + emit_scores)
-        else:
-            alpha = torch.exp(emit_scores)
-
-        # z.shape = (batch_size, 1)
-        z = alpha.sum(dim=1, keepdim=True)
-        alpha = alpha / z
-        sum_log_z = torch.log(z) * label_weights[tags[0]]
-
-        # For each i we compute logits for the transitions from timestep i-1 to timestep i.
-        # We do so in a (batch_size, num_tags, num_tags) tensor where the axes are
-        # (instance, current_tag, next_tag)
-        for i in range(1, sequence_length):
-            # multiply the logits by the label weights
-            # logits[i].shape: (batch_size, num_tags)
-            # emit_scores = torch.mul(logits[i], label_weights)
-            emit_scores = logits[i]
-
-            # The emit scores are for time i ("next_tag") so we broadcast along the current_tag axis.
-            emit_scores = emit_scores.view(batch_size, 1, num_tags)
-            # Transition scores are (current_tag, next_tag) so we broadcast along the instance axis.
-            transition_scores = self.transitions.view(1, num_tags, num_tags)
-            # Alpha is for the current_tag (i-1), so we broadcast along the next_tag axis.
-            broadcast_alpha = alpha.view(batch_size, num_tags, 1)
-
-            # Add all the scores together and logexp over the current_tag axis.
-            inner = broadcast_alpha * torch.exp(emit_scores + transition_scores)
-
-            # In valid positions (mask == True) we want to take the logsumexp over the current_tag dimension
-            # of `inner`. Otherwise (mask == False) we want to retain the previous alpha.
-            alpha = inner.sum(dim=1) * mask[i].view(batch_size, 1) + alpha * (~mask[i]).view(
-                batch_size, 1
-            )
-
-            # scale alphas to avoid underflow (sum of alphas equal to 1)
-            z = alpha.sum(dim=1, keepdim=True)
-            alpha = alpha / z
-            # weight z (normalization factor) according to the current tag
-            sum_log_z += torch.log(z) * label_weights[tags[i]]
-
-        # Every sequence needs to end with a transition to the stop_tag.
-        if self.include_start_end_transitions:
-            alpha = alpha * torch.exp(self.end_transitions.view(1, num_tags))
-            z = alpha.sum(dim=1, keepdim=True)
-            # alpha = alpha / z  # this step is unnecessary since alpha is not used anymore
-            sum_log_z += torch.log(z)
-
-        return sum_log_z.squeeze(1)
-
-    def _joint_likelihood(
-        self, logits: torch.Tensor, tags: torch.Tensor, mask: torch.BoolTensor
-    ) -> torch.Tensor:
-        """
-        Computes the numerator term for the log-likelihood, which is just score(inputs, tags)
-        """
-        batch_size, sequence_length, _ = logits.data.shape
-
-        # Transpose batch size and sequence dimensions:
-        logits = logits.transpose(0, 1).contiguous()
-        mask = mask.transpose(0, 1).contiguous()
-        tags = tags.transpose(0, 1).contiguous()
-
-        # Start with the transition scores from start_tag to the first tag in each input
-        if self.include_start_end_transitions:
-            score = self.start_transitions.index_select(0, tags[0])
-        else:
-            score = 0.0
-
-        label_weights = self.label_weights
-
-        # weight transition score using current_tag, i.e., t(i,j) will be t(i,j)*w(i),
-        # where t(i,j) is the score to transition from i to j and w(i) is the weight
-        # for tag i.
-        transitions = self.transitions * label_weights.view(-1, 1)
-
-        # Add up the scores for the observed transitions and all the inputs but the last
-        for i in range(sequence_length - 1):
-            # Each is shape (batch_size,)
-            current_tag, next_tag = tags[i], tags[i + 1]
-
-            # The scores for transitioning from current_tag to next_tag
-            transition_score = transitions[current_tag.view(-1), next_tag.view(-1)]
-
-            # The score for using current_tag
-            emit_score = logits[i].gather(1, current_tag.view(batch_size, 1)).squeeze(1)
-
-            # Weight emit scores by label.
-            emit_score *= label_weights[current_tag.view(-1)]
-
-            # Include transition score if next element is unmasked,
-            # input_score if this element is unmasked.
-            score = score + transition_score * mask[i + 1] + emit_score * mask[i]
-
-        # Transition from last state to "stop" state. To start with, we need to find the last tag
-        # for each instance.
-        last_tag_index = mask.sum(0).long() - 1
-        last_tags = tags.gather(0, last_tag_index.view(1, batch_size)).squeeze(0)
-
-        # Compute score of transitioning to `stop_tag` from each "last tag".
-        if self.include_start_end_transitions:
-            last_transition_score = self.end_transitions.index_select(0, last_tags)
-        else:
-            last_transition_score = 0.0
-
-        # Add the last input if it's not masked.
-        last_inputs = logits[-1]  # (batch_size, num_tags)
-        last_input_score = last_inputs.gather(1, last_tags.view(-1, 1))  # (batch_size, 1)
-        last_input_score = last_input_score.squeeze()  # (batch_size,)
-
-        # Weight last emit scores by label weights.
-        last_input_score = last_input_score * label_weights[last_tags.view(-1)]
-
-        score = score + last_transition_score + last_input_score * mask[-1]
-
-        return score
-
-    def forward(
-        self, inputs: torch.Tensor, tags: torch.Tensor, mask: torch.BoolTensor = None
-    ) -> torch.Tensor:
-        """
-        Computes the log likelihood.
-        """
-
-        if mask is None:
-            mask = torch.ones(*tags.size(), dtype=torch.bool, device=inputs.device)
-        else:
-            # The code below fails in weird ways if this isn't a bool tensor, so we make sure.
-            mask = mask.to(torch.bool)
-
-        # TODO check if weights are being used during test/validation
-
-        log_denominator = self._input_likelihood(inputs, tags, mask)
-        log_numerator = self._joint_likelihood(inputs, tags, mask)
-
-        return torch.sum(log_numerator - log_denominator)
-
-    def viterbi_tags(
-        self, logits: torch.Tensor, mask: torch.BoolTensor = None, top_k: int = None
-    ) -> Union[List[VITERBI_DECODING], List[List[VITERBI_DECODING]]]:
-        """
-        Uses viterbi algorithm to find most likely tags for the given inputs.
-        If constraints are applied, disallows all other transitions.
-
-        Returns a list of results, of the same size as the batch (one result per batch member)
-        Each result is a List of length top_k, containing the top K viterbi decodings
-        Each decoding is a tuple  (tag_sequence, viterbi_score)
-
-        For backwards compatibility, if top_k is None, then instead returns a flat list of
-        tag sequences (the top tag sequence for each batch item).
-        """
-        if mask is None:
-            mask = torch.ones(*logits.shape[:2], dtype=torch.bool, device=logits.device)
-
-        if top_k is None:
-            top_k = 1
-            flatten_output = True
-        else:
-            flatten_output = False
-
-        _, max_seq_length, num_tags = logits.size()
-
-        # Get the tensors out of the variables
-        logits, mask = logits.data, mask.data
-
-        # Augment transitions matrix with start and end transitions
-        start_tag = num_tags
-        end_tag = num_tags + 1
-        transitions = torch.full((num_tags + 2, num_tags + 2), -10000.0, device=logits.device)
-
-        # Apply transition constraints
-        constrained_transitions = self.transitions * self._constraint_mask[
-            :num_tags, :num_tags
-        ] + -10000.0 * (1 - self._constraint_mask[:num_tags, :num_tags])
-        transitions[:num_tags, :num_tags] = constrained_transitions.data
-
-        if self.include_start_end_transitions:
-            transitions[
-                start_tag, :num_tags
-            ] = self.start_transitions.detach() * self._constraint_mask[
-                start_tag, :num_tags
-            ].data + -10000.0 * (
-                1 - self._constraint_mask[start_tag, :num_tags].detach()
-            )
-            transitions[:num_tags, end_tag] = self.end_transitions.detach() * self._constraint_mask[
-                :num_tags, end_tag
-            ].data + -10000.0 * (1 - self._constraint_mask[:num_tags, end_tag].detach())
-        else:
-            transitions[start_tag, :num_tags] = -10000.0 * (
-                1 - self._constraint_mask[start_tag, :num_tags].detach()
-            )
-            transitions[:num_tags, end_tag] = -10000.0 * (
-                1 - self._constraint_mask[:num_tags, end_tag].detach()
-            )
-
-        best_paths = []
-        # Pad the max sequence length by 2 to account for start_tag + end_tag.
-        tag_sequence = torch.empty(max_seq_length + 2, num_tags + 2, device=logits.device)
-
-        for prediction, prediction_mask in zip(logits, mask):
-            mask_indices = prediction_mask.nonzero(as_tuple=False).squeeze()
-            masked_prediction = torch.index_select(prediction, 0, mask_indices)
-            sequence_length = masked_prediction.shape[0]
-
-            # Start with everything totally unlikely
-            tag_sequence.fill_(-10000.0)
-            # At timestep 0 we must have the START_TAG
-            tag_sequence[0, start_tag] = 0.0
-            # At steps 1, ..., sequence_length we just use the incoming prediction
-            tag_sequence[1 : (sequence_length + 1), :num_tags] = masked_prediction
-            # And at the last timestep we must have the END_TAG
-            tag_sequence[sequence_length + 1, end_tag] = 0.0
-
-            # We pass the tags and the transitions to `viterbi_decode`.
-            viterbi_paths, viterbi_scores = util.viterbi_decode(
-                tag_sequence=tag_sequence[: (sequence_length + 2)],
-                transition_matrix=transitions,
-                top_k=top_k,
-            )
-            top_k_paths = []
-            for viterbi_path, viterbi_score in zip(viterbi_paths, viterbi_scores):
-                # Get rid of START and END sentinels and append.
-                viterbi_path = viterbi_path[1:-1]
-                top_k_paths.append((viterbi_path, viterbi_score.item()))
-            best_paths.append(top_k_paths)
-
-        if flatten_output:
-            return [top_k_paths[0] for top_k_paths in best_paths]
-
-        return best_paths
