+++ allennlp/modules/conditional_random_field_weighted/__init__.py
@@ -1,0 +1,3 @@
+from allennlp.modules.conditional_random_field_weighted.conditional_random_field_wemission import ConditionalRandomFieldWeightEmission
+from allennlp.modules.conditional_random_field_weighted.conditional_random_field_wtrans import ConditionalRandomFieldWeightTrans
+from allennlp.modules.conditional_random_field_weighted.conditional_random_field_wlannoy import ConditionalRandomFieldWeightLannoy
